refactor: refactor qdp-python `lib.rs`
